requirements.txt
streamlit
pillow
easyocr
torch
typing_extensions